Replace NativeList int64 indexing with int32
Alongside a new MAX_CAPACITY variable, this ensures lists cannot exceed the byte size limits of a span.

+++ YARG.Core/NewParsing/Sync/TempoTracker.cs
@@ -10,7 +10,7 @@ namespace YARG.Core.NewParsing
     {
         private readonly long                               _resolution;
         private readonly YargNativeSortedList<long, Tempo2> _tempoMarkers;
-        private          long                               _index;
+        private          int                                _index;
 
         public readonly long Resolution => _resolution;
